# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Organization service for create/update/delete organization for user. It allows to invite/remove other users, assign them role and update their roles.",
        "title": "Organization-Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000/",
    "paths": {
...
        "/user/organizations/": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch all organizations of current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserOrganizationData"
                ],
                "summary": "fetch all organizations",
                "operationId": "organizations-of-user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AllOrganizationDetailsOfUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
        "/user/organizations/{organization-id}": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch organization details of current user by organizationID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserOrganizationData"
                ],
                "summary": "fetch organization details of current user",
                "operationId": "organization-of-user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OrganizationDetailsOfUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "error.CustomError": {
            "type": "object",
            "required": [
                "errorMessage",
                "statusCode"
            ],
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "invalidData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/error.InvalidData"
                    }
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "error.InvalidData": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "field": {
                    "type": "string"
                }
            }
        },
...
        "response.AllOrganizationDetailsOfUser": {
            "type": "object",
            "properties": {
                "organizations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Organization"
                    }
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "response.Organization": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizationID": {
                    "type": "string"
                },
                "organizationMembers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.OrganizationMember"
                    }
                },
                "ownerID": {
                    "type": "string"
                },
                "privacy": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "response.OrganizationDetailsOfUser": {
            "type": "object",
            "properties": {
                "organization": {
                    "$ref": "#/definitions/response.Organization"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "response.OrganizationMember": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Auth-user",
            "in": "header"
        },
        "jwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}